refactor: SSR / CSR 컴포넌트 분리-#132
