+++ omechu-app/src/app/components/common/RestaurantReviewCard.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @next/next/no-img-element */
 import { useState } from "react";
 
 import Image from "next/image";
@@ -62,7 +63,7 @@ export default function RestaurantReviewCard({
     const [imgSrc, setImgSrc] = useState(src || "/image/image_empty.svg");
 
     return (
-      <Image
+      <img
         src={imgSrc}
         alt={alt}
         width={70}
@@ -84,10 +85,9 @@ export default function RestaurantReviewCard({
 
     return (
       <div className="relative h-[100px] min-w-[100px] flex-shrink-0">
-        <Image
+        <img
           src={imgSrc}
           alt={alt}
-          fill
           className="object-cover"
           onError={() => setImgSrc("/image/image_empty.svg")}
         />
@@ -115,7 +115,7 @@ export default function RestaurantReviewCard({
                 <span className="text-lg font-normal leading-tight text-grey-darker">
                   {restaurantName}
                   <button onClick={onNavigate}>
-                    <Image
+                    <img
                       src="/arrow/right_arrow_navigate_next.svg"
                       alt={"이동 버튼"}
                       width={25}
@@ -138,7 +138,7 @@ export default function RestaurantReviewCard({
                 {recommendCount ?? 0}
               </span>
               <button onClick={onLikeToggle}>
-                <Image
+                <img
                   src={
                     isLiked
                       ? "/thumb/thumbs-up-fill.svg"
